refactor: cardStyles를 lib 레이어로 이동
cardStyles는 React가 필요 없는 순수 문자열 상수이므로 lib 레이어로 이동하여 레이어 의존성 정확히 함
