+++ src/components/dashboard/IndexCard.tsx
@@ -1,5 +1,5 @@
 import Link from 'next/link';
-import { CARD_LINK_CLASSES } from '@/components/ui/cardStyles';
+import { CARD_LINK_CLASSES } from '@/lib/cardStyles';
 import { cn } from '@/lib/cn';
 import { formatPriceChange, formatUsdPrice } from '@/lib/priceFormat';
 import type { MarketIndexData, MarketSectorData } from '@/domain/types';
